+++ common/lib/src/realm_types.dart
@@ -79,6 +79,29 @@ class RealmStateError extends StateError implements RealmError {
   RealmStateError(String message) : super(message);
 }
 
+class SessionError extends SyncError {
+  final bool isFatal;
+
+  SessionError(
+    String message,
+    this.isFatal, {
+    SyncErrorCategory category = SyncErrorCategory.unknown,
+    int code = 0,
+  }) : super(message, category, code);
+}
+
+/// Thrown or reporeted if an error occurs during synchronization
+/// {@category Sync}
+class SyncError extends RealmError {
+  /// The code of the error
+  final int code; // TODO: this should be an enum
+
+  /// The category of the error
+  final SyncErrorCategory category;
+
+  SyncError(String message, this.category, this.code) : super(message);
+}
+
 /// The category of a [SyncError].
 enum SyncErrorCategory {
   /// The error originated from the client
@@ -95,29 +118,6 @@ enum SyncErrorCategory {
 
   /// The category is unknown
   unknown,
-}
-
-/// Thrown or reporeted if an error occurs during synchronization
-/// {@category Sync}
-class SyncError extends RealmError {
-  /// The code of the error
-  final int code; // TODO: this should be an enum
-
-  /// The category of the error
-  final SyncErrorCategory category;
-
-  SyncError(String message, this.category, this.code) : super(message);
-}
-
-class SessionError extends SyncError {
-  final bool isFatal;
-
-  SessionError(
-    String message,
-    this.isFatal, {
-    SyncErrorCategory category = SyncErrorCategory.unknown,
-    int code = 0,
-  }) : super(message, category, code);
 }
 
 /// @nodoc